backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
resend
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
httpx
pydantic-settings